--- 3/index3.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>Force-directed Graph with Directional Straight Arrow</title>
-    <link rel="stylesheet" href="style.css" />
-  </head>
-  <body>
-    <svg id="graph"></svg>
-</body>
-<script src="https://d3js.org/d3.v6.min.js"></script>
-<script src="script.js"></script>
-</html>
