--- 3/index3.md
@@ -1,6 +1,0 @@
-
-
-ChatGPT inspiré de :
-- https://observablehq.com/@brunolaranjeira/d3-v6-force-directed-graph-with-directional-straight-arrow
-
-
