--- 3/style.css
@@ -1,15 +1,0 @@
-.node {
-    fill: #333;
-    stroke: #fff;
-    stroke-width: 1.5px;
-  }
-  .link {
-    stroke: #999;
-    stroke-opacity: 0.6;
-    marker-end: url(#arrowhead);
-  }
-  .label {
-    font-size: 12px;
-    font-family: sans-serif;
-    fill: #aaa;
-  }
