--- 4/date.json
@@ -1,16 +1,0 @@
-{
-    "nodes": [
-      { "id": "0", "name": "Node 0" },
-      { "id": "1", "name": "Node 1" },
-      { "id": "2", "name": "Node 2" },
-      { "id": "3", "name": "Node 3" },
-      { "id": "4", "name": "Node 4" }
-    ],
-    "links": [
-      { "source": "0", "target": "1", "value": 1 },
-      { "source": "1", "target": "2", "value": 1 },
-      { "source": "2", "target": "3", "value": 1 },
-      { "source": "3", "target": "4", "value": 1 },
-      { "source": "4", "target": "0", "value": 1 }
-    ]
-}
