--- 9/json/graph9.json
@@ -1,90 +1,0 @@
-{
-    "nodes": [
-        {
-            "id": "1",
-            "nom": "Node 1",
-            "description": "Description 1",
-            "index": 0,
-            "x": 274.309407793824,
-            "y": 231.6084659539223,
-            "vy": -0.0003242174514073547,
-            "vx": 0.0011898074234974831
-        },
-        {
-            "id": "2",
-            "nom": "Node 2",
-            "description": "Description 2",
-            "index": 1,
-            "x": 373.46540665557296,
-            "y": 248.86251579035516,
-            "vy": -0.001413148490322341,
-            "vx": 0.000743146118118971
-        },
-        {
-            "id": "3",
-            "nom": "Node 3",
-            "description": "Description 3",
-            "index": 2,
-            "x": 474.25857012879635,
-            "y": 236.21143023218065,
-            "vy": 0.0009653046239788772,
-            "vx": -0.0006359600147902029
-        },
-        {
-            "id": "1683778789343",
-            "nom": "Node 1683778789343",
-            "description": "Description 1683778789343",
-            "x": 520.0697360701203,
-            "y": 326.21586821044326,
-            "index": 3,
-            "vy": -0.00017140066479314782,
-            "vx": 0.0004942804277250637
-        },
-        {
-            "id": "1683778789810",
-            "nom": "Node 1683778789810",
-            "description": "Description 1683778789810",
-            "x": 330.1529842449254,
-            "y": 390.51675140017693,
-            "index": 4,
-            "vy": 0.0008631853863614762,
-            "vx": 0.0008967649251014964
-        },
-        {
-            "id": "1683778791466",
-            "nom": "Node 1683778791466",
-            "description": "Description 1683778791466",
-            "x": 427.74745060882304,
-            "y": 366.5856284056897,
-            "index": 5,
-            "vy": 0.0007402693640643051,
-            "vx": 0.0008674631826630172
-        }
-    ],
-    "links": [
-        {
-            "source": "1",
-            "target": "2"
-        },
-        {
-            "source": "2",
-            "target": "3"
-        },
-        {
-            "source": "1683778789810",
-            "target": "1683778791466"
-        },
-        {
-            "source": "1683778789343",
-            "target": "1683778789343"
-        },
-        {
-            "source": "3",
-            "target": "1683778789343"
-        },
-        {
-            "source": "1683778791466",
-            "target": "1683778789343"
-        }
-    ]
-}
