--- 9/json_config/force_field.json
@@ -1,1 +1,0 @@
-{"linkStrength":1,"linkDistance":200,"chargeStrength":-50,"centerStrength":1}
