--- 9/json_config/manual_without_force.json
@@ -1,1 +1,0 @@
-{"linkStrength":0,"linkDistance":200,"chargeStrength":0,"centerStrength":0}
