--- index.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="UTF-8">
-  <title>Network Graph with d3.js v6</title>
-  <link rel="stylesheet" href="css/style.css" />
-</head>
-<body>
-  <g class="labels"></g>
-  <g class="node-labels"></g>
-<g class="node-labels"></g>
-
-  <div id="graph"></div>
-  <div id="config-menu">
-    <!-- Ajoutez ici les éléments du menu de configuration -->
-  </div>
-  <!-- Import/export 
-  <button id="importButton">Import JSON</button>
-  <input type="file" id="importFile" style="display:none">
-  <button id="exportButton">Export JSON</button>
-  -->
-
-</body>
-<script src="https://d3js.org/d3.v6.min.js"></script>
-<!-- Important, insérer le script après le body pour que ça marche -->
-<script src="js/graph.js"></script>
-</html>
