--- index2.html
@@ -1,67 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Simple Network Graph</title>
-  <script src="https://d3js.org/d3.v6.min.js"></script>
-  <style>
-    body { margin: 0; font-family: sans-serif; }
-    svg { display: block; margin: auto; }
-  </style>
-</head>
-<body>
-  <script>
-    // Test d'un graph simple avec d3.js v6
-    const width = 800;
-    const height = 600;
-
-    const svg = d3.select("body")
-      .append("svg")
-      .attr("width", width)
-      .attr("height", height);
-
-    const nodesData = [
-      {id: 1},
-      {id: 2},
-      {id: 3}
-    ];
-
-    const linksData = [
-      {source: 1, target: 2},
-      {source: 2, target: 3}
-    ];
-
-    const simulation = d3.forceSimulation(nodesData)
-      .force("link", d3.forceLink(linksData).id(d => d.id))
-      .force("charge", d3.forceManyBody())
-      .force("center", d3.forceCenter(width / 2, height / 2));
-
-    const link = svg.append("g")
-      .attr("stroke", "#999")
-      .attr("stroke-opacity", 0.6)
-      .selectAll("line")
-      .data(linksData)
-      .join("line");
-
-    const node = svg.append("g")
-      .attr("fill", "#4B92DB")
-      .selectAll("circle")
-      .data(nodesData)
-      .join("circle")
-      .attr("r", 7);
-
-    simulation.on("tick", () => {
-      link
-        .attr("x1", d => d.source.x)
-        .attr("y1", d => d.source.y)
-        .attr("x2", d => d.target.x)
-        .attr("y2", d => d.target.y);
-
-      node
-        .attr("cx", d => d.x)
-        .attr("cy", d => d.y);
-    });
-  </script>
-</body>
-</html>
