+++ index9.html
@@ -10,7 +10,7 @@
   <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.16.0/umd/popper.min.js"></script>
   <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
 
-  <link rel="stylesheet" href="style.css" />
+  <link rel="stylesheet" href="css/style.css" />
 </head>
 <body>
   <svg width="800" height="600">
@@ -121,6 +121,10 @@
                 <label for="link-label">Label du lien:</label>
                 <select id="link-label">
                 </select>
+                <label for="node-size">Node size : </label>
+                <select id="node-size"></select>
+                <input type="number" id="node-size-input" value="10" step="0.1" min="0" max="100" onchange="updateNodeSize()">
+
               </p>
               <hr>
               <form id="node-form" class="hidden"></form>
